# docs/api/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Now Shipping Public API
  version: 1.0.0
  description: |
    Public REST API for business integrations with Now Shipping.
    Fees are always computed server-side — never send orderFees or pickupFees in requests.
    Company integrators use one API key and send X-Merchant-Id per order/pickup request.
  contact:
    name: Now Shipping Integrations
servers:
  - url: /api/public/v1
    description: Public API v1
security:
  - BearerAuth: []
  - ApiKeyHeader: []
...
paths:
...
  /pickups:
    get:
      tags: [Pickups]
      summary: List pickups
      parameters:
        - $ref: '#/components/parameters/MerchantId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Pickup list
    post:
      tags: [Pickups]
      summary: Create pickup
      parameters:
        - $ref: '#/components/parameters/MerchantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PickupCreateRequest'
      responses:
        '201':
          description: Pickup created
...
  /orders:
    get:
      tags: [Orders]
      summary: List orders
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: orderType
          in: query
          schema:
            type: string
            enum: [Deliver, Return, Exchange, All]
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessOrderList'
    post:
      tags: [Orders]
      summary: Create order
      description: Fees are computed automatically. Do not send orderFees.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessOrderCreate'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /orders/{orderId}/cancel:
    post:
      tags: [Orders]
      summary: Cancel order
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Canceled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key as Bearer token (nsk_live_...)
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
    MerchantId:
      name: X-Merchant-Id
      in: header
      description: Merchant businessAccountCode, externalMerchantId, or Mongo _id (required for company API keys on order/pickup routes)
      schema:
        type: string
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
      description: MongoDB _id or order number
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    OrderCreateRequest:
      type: object
      required: [fullName, phoneNumber, address, government, zone, orderType]
      properties:
        fullName:
          type: string
        phoneNumber:
          type: string
        otherPhoneNumber:
          type: string
        address:
          type: string
        buildingNo:
          type: string
        apartmentNo:
          type: string
        government:
          type: string
          enum: [Cairo, Giza, Qalyubia]
        zone:
          type: string
        deliverToWorkAddress:
          type: boolean
        orderType:
          type: string
          enum: [Deliver, Return, Exchange]
        productDescription:
          type: string
        numberOfItems:
          type: integer
        isExpressShipping:
          type: boolean
        COD:
          type: boolean
        amountCOD:
          type: number
        CashDifference:
          type: boolean
        amountCashDifference:
          type: number
        previewPermission:
          type: boolean
        referralNumber:
          type: string
        Notes:
          type: string
        originalOrderNumber:
          type: string
        selectedPickupAddressId:
          type: string
...
    PickupCreateRequest:
      type: object
      required: [numberOfOrders, pickupDate, phoneNumber]
      properties:
        numberOfOrders:
          type: integer
        pickupDate:
          type: string
          format: date
        phoneNumber:
          type: string
        isFragileItems:
          type: boolean
        isLargeItems:
          type: boolean
        pickupNotes:
          type: string
        pickupLocation:
          type: string
        pickupAddressId:
          type: string
...
    OrderSummary:
      type: object
      properties:
        orderId:
          type: string
        orderNumber:
          type: string
        orderStatus:
          type: string
        orderFees:
          type: number
        orderType:
          type: string
        isExpressShipping:
          type: boolean
    SuccessOrderCreate:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            message:
              type: string
            order:
              $ref: '#/components/schemas/OrderSummary'
...
    SuccessOrderList:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            orders:
              type: array
              items:
                $ref: '#/components/schemas/OrderSummary'
            pagination:
              type: object
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            message:
              type: string