# routes/web/adminRoutes.js
const express = require('express');
const router = express.Router();
const jwt = require('jsonwebtoken');
const Admin = require('../../models/admin.js');
const Notification = require('../../models/notification.js');

const adminController = require('../../controllers/adminController.js');
const applicationController = require('../../controllers/applicationController.js');
const adminWhatsAppController = require('../../controllers/adminWhatsAppController.js');
const notificationController = require('../../controllers/notificationController.js');
const apiTokenController = require('../../controllers/apiTokenController.js');
const adminMerchantController = require('../../controllers/adminMerchantController.js');

const jwtSecret = process.env.JWT_SECRET;

async function authenticateAdmin(req, res, next) {
  const token = req.cookies.token;
  if (!token) {
    console.log('Token not found');
    return res.status(401).redirect('/admin-login');
  }

  try {
    const decode = jwt.verify(token, jwtSecret);
    req.adminId = decode.adminId;
    const admin = await Admin.findOne({ _id: decode.adminId });
    if (!admin) {
      res.clearCookie('token');
      return res.status(401).redirect('/admin-login');
    }
    req.adminData = admin; // Attach admin data to request object

    // Make adminData available to all views
    res.locals.adminData = admin;

    next(); // Move to the next middleware
  } catch (error) {
    res.clearCookie('token');
    return res.status(401).redirect('/admin-login');
  }
}

router.use(authenticateAdmin);

router.get('/dashboard', adminController.getDashboardPage);
router.get('/dashboard-data', adminController.getAdminDashboardData);
router.get('/get-delivery-men', adminController.get_deliveryMenByZone);
router.post('/assign-delivery-man', adminController.assignCourierToStock);

router.get('/orders', adminController.get_ordersPage);
router.get('/get-orders', adminController.get_orders);
router.get(
  '/orders-filter-businesses',
  adminController.get_ordersFilterBusinesses
);
router.get('/order-details/:orderNumber', adminController.get_orderDetailsPage);
router.get('/print-policy/:orderNumber', adminController.printPolicy);
router.post(
  '/orders/print-policy/:orderNumber/:pageSize',
  adminController.printPolicy
);

// couriers

router.get('/couriers', adminController.get_couriersPage);

router.get('/get-couriers', adminController.get_couriers);

router.post('/couriers/create-courier', adminController.createCourier);

router.put('/couriers/update-zones', adminController.updateCourierZones);

router.post('/couriers/bulk-deactivate', adminController.bulkDeactivateCouriers);
router.post('/couriers/bulk-delete', adminController.bulkDeleteCouriers);

// couriers-follow-up

router.get('/couriers-follow-up', adminController.get_couriersFollowUp);

// router.get('/get-couriers-follow-up', adminController.get_couriersFollowUp);

router.get(
  '/courier-details/:courierId',
  adminController.get_courierDetailsPage
);

// pickups
router.get('/pickups', adminController.get_pickupsPage);

router.get('/get-pickups', adminController.get_pickups);

router.get(
  '/pickup-details/:pickupNumber',
  adminController.get_pickupDetailsPage
);

router.get(
  '/pickup-details/:pickupNumber/get-pickedup-orders',
  adminController.get_pickedupOrders
);

router.put('/cancel-pickup/:pickupId', adminController.cancelPickup);

router.delete('/delete-pickup/:pickupId', adminController.deletePickup);

router.get('/get-pickup-men', adminController.get_pickupMenByZone);

router.post('/assign-pickup-man', adminController.assignPickupMan);

//======= stock management ==========//

// stock Managment Pickup

router.get(
  '/stock-management/pickups',
  adminController.get_stockManagementPage
);

router.get('/get-stock-orders', adminController.get_stock_orders);

router.post('/add-to-stock', adminController.add_to_stock);

router.get('/get-couriers-by-zone', adminController.get_couriers_by_zone);

router.post(
  '/stock-managment/assign-courier',
  adminController.assignCourierToStock
);

router.post(
  '/stock-managment/courier-received',
  adminController.courier_received
);

// stock Managment Returns

router.get('/stock-management/returns', adminController.get_stockReturnsPage);

router.get('/get-return-orders', adminController.getReturnedOrders);

router.post('/add-return-to-stock', adminController.add_return_to_stock);

router.post('/return-assign-courier', adminController.assignCourierToReturn);

// Resend return pickup OTP to customer (if expired or not received)
router.post('/orders/:orderNumber/resend-return-otp', adminController.resendReturnOtp);

router.post(
  '/return-courier-received',
  adminController.return_courier_received
);

// return management
router.post(
  '/assign-courier-to-return-to-business',
  adminController.assignCourierToReturnToBusiness
);

// Enhanced Return Flow routes
router.post(
  '/orders/:orderId/convert-to-return/:reason',
  adminController.convertFailedDeliveryToReturn
);
router.get('/return-orders', adminController.getAllReturnOrders);
router.get(
  '/return-orders/:orderId',
  adminController.getReturnOrderDetailsAdmin
);
router.put(
  '/return-orders/:orderId/inspection',
  adminController.updateReturnInspection
);
router.put(
  '/return-orders/:orderId/processing',
  adminController.updateReturnProcessing
);

// router.get('/get-stock-managment', adminController.get_stockManagment);

// Payouts
router.get('/payouts', adminController.get_payoutsPage);
router.get('/api/businesses/search', adminController.search_businessesForPayout);
router.get('/api/payouts', adminController.get_payouts);
router.get('/api/payouts/:id', adminController.get_payoutDetail);
router.get('/api/payouts/:id/export', adminController.export_payoutExcel);
router.post('/api/payouts/:id/paid', adminController.markPayoutPaid);
router.post('/api/adjustments', adminController.addAdjustment);
router.get('/business/:id/ledger', adminController.get_businessLedgerPage);
router.get('/api/business/:id/ledger', adminController.get_businessLedger);
router.post('/api/run-payout-processing', adminController.runPayoutProcessing);

// ── TEMP: One-time data migration (remove after running once) ──
router.post('/api/migrate-to-ledger', async (req, res) => {
  try {
    const { migrateToLedger } = require('../utils/migrateToLedger');
    const dryRun = req.query.dryRun === 'true';
    const summary = await migrateToLedger({ dryRun });
    res.json({ success: true, summary });
  } catch (err) {
    res.status(500).json({ success: false, error: err.message });
  }
});

// businesses

router.get('/businesses', adminController.get_businessesPage);
router.get('/get-businesses', adminController.get_businesses);
router.get(
  '/business-details/:businessId',
  adminController.get_businessDetailsPage
);
router.get(
  '/get-business-details/:businessId',
  adminController.get_businessDetails
);
router.get('/api/business/:id/pricing', adminController.get_businessPricing);
router.put('/api/business/:id/pricing', adminController.update_businessPricing);
router.get(
  '/business/:businessId/deletion-impact',
  adminController.get_businessDeletionImpact
);
router.post('/business/:businessId/delete', adminController.delete_business);

// Business API tokens (public integration keys)
router.get('/business/:businessId/api-tokens', apiTokenController.listApiTokens);
router.post('/business/:businessId/api-tokens', apiTokenController.createApiToken);
router.post('/business/:businessId/api-tokens/:tokenId/revoke', apiTokenController.revokeApiToken);

// Multi-tenant company / merchant management
router.put('/business/:businessId/account-type', adminMerchantController.updateAccountType);
router.put('/business/:businessId/parent-company', adminMerchantController.updateParentCompany);
router.get('/business/:businessId/merchants', adminMerchantController.listCompanyMerchants);
router.get('/api/companies/search', adminMerchantController.searchCompanies);

// tickets
router.get('/tickets', adminController.get_ticketsPage);

// landing partner applications
router.get('/applications', applicationController.getApplicationsPage);
router.get('/get-applications', applicationController.getApplications);
router.get('/applications/:id', applicationController.getApplicationById);
router.patch('/applications/:id', applicationController.updateApplication);

//logout
router.get('/logout', adminController.logOut);

// waitingAction admin overrides
router.post(
  '/orders/:orderId/retry-tomorrow',
  adminController.adminRetryTomorrow
);
router.post(
  '/orders/:orderId/retry-scheduled',
  adminController.adminRetryScheduled
);
router.post(
  '/orders/:orderId/return-to-warehouse',
  adminController.adminReturnToWarehouseFromWaiting
);
router.post('/orders/cancel-order/:orderId', adminController.adminCancelOrder);
router.post('/orders/complete-order/:orderId', adminController.adminCompleteOrder);
router.post('/orders/complete-multiple', adminController.adminCompleteMultiple);
router.delete('/orders/delete-order/:orderId', adminController.adminDeleteOrder);
router.post('/orders/delete-multiple', adminController.adminDeleteMultiple);
router.get('/edit-order/:orderNumber', adminController.get_adminEditOrderPage);
router.put('/orders/edit-order/:orderId', adminController.adminEditOrder);
router.post('/orders/:orderId/cancel', adminController.adminCancelFromWaiting);
router.post(
  '/orders/:orderId/change-return-courier',
  adminController.changeReturnCourier
);

// Courier Tracking
router.get('/courier-tracking', adminController.courierTracking);
router.get('/courier-locations', adminController.getCourierLocations);
router.get('/courier-location/:id', adminController.getCourierLocation);

// Notification routes - Couriers
router.get('/notifications', notificationController.getNotificationsPage);
router.post(
  '/notifications/courier',
  notificationController.sendNotificationToCourier
);
router.post(
  '/notifications/broadcast',
  notificationController.sendNotificationToAllCouriers
);

// Notification routes - Businesses
router.get('/notifications/businesses', notificationController.getBusinessNotificationsPage);
router.post(
  '/notifications/business',
  notificationController.sendNotificationToBusiness
);
router.post(
  '/notifications/businesses/broadcast',
  notificationController.sendNotificationToAllBusinesses
);

// Shared notification routes
router.get(
  '/notifications/recent',
  notificationController.getRecentNotifications
);

// FCM Token Management routes
router.post(
  '/notifications/cleanup-tokens',
  notificationController.cleanupInvalidTokens
);
router.post('/notifications/test-token', notificationController.testUserToken);

// Admin Shop API routes (must come before page routes to avoid conflicts)
router.get('/api/shop/products', adminController.getProducts);
router.post('/api/shop/products', adminController.createProduct);
router.get('/api/shop/products/:id', adminController.getProduct);
router.put('/api/shop/products/:id', adminController.updateProduct);
router.delete('/api/shop/products/:id', adminController.deleteProduct);

router.get('/api/shop/orders', adminController.getShopOrders);
router.get('/api/shop/orders/:id', adminController.getShopOrder);
router.put(
  '/api/shop/orders/:id/status',
  adminController.updateShopOrderStatus
);
router.put(
  '/api/shop/orders/:id/assign-courier',
  adminController.assignCourierToShopOrder
);
router.post(
  '/api/shop/orders/assign-multiple-couriers',
  adminController.assignMultipleCouriersToShopOrders
);
router.get('/api/shop/couriers-by-zone', adminController.get_deliveryMenByZone);
router.get('/api/couriers', adminController.get_couriers);
router.get('/api/admin/couriers', adminController.getAllCouriers);

// Shop Management page routes
router.get('/shop/products', adminController.getShopProductsPage);
router.get('/shop/orders', adminController.getShopOrdersPage);
router.get('/shop/orders/:id', adminController.getShopOrderDetailsPage);

// Smart Flyers Management
router.get('/print-smart-flyers', adminController.getPrintSmartFlyersPage);
router.post('/generate-smart-flyers', adminController.generateSmartFlyers);

// Wasender: WhatsApp sessions & QR (personal access token)
router.get('/connect-whatsapp', adminWhatsAppController.getConnectWhatsAppPage);
router.get('/api/whatsapp-sessions', adminWhatsAppController.apiListSessions);
router.post(
  '/api/whatsapp-sessions/:sessionId/connect',
  adminWhatsAppController.apiConnectSession
);
router.get(
  '/api/whatsapp-sessions/:sessionId/qrcode',
  adminWhatsAppController.apiGetQr
);

module.exports = router;